requirements.txt
pyyaml
boto3==1.9.112
